refactor(gateway): extract slash command handlers
